refactor: split faster_geometry into package
